refactor find related classes
